requirements.txt
dj-database-url==0.4.1
appdirs==1.4.3
gunicorn==19.7.1
packaging==16.8
pyparsing==2.2.0
six==1.10.0
whitenoise==3.3.0
psycopg2==2.7.3.2
psycopg2-binary
certifi==2017.4.17
chardet==3.0.4
Django==1.11.2
idna==2.5
pycountry==17.5.14
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
django-htmlmin==0.10.0




